RAG-Based Customer Support Assistant - Final Project/requirements.txt
langchain==0.1.0
chromadb==0.4.15
tiktoken==0.5.1
python-dotenv==1.0.0
streamlit==1.28.1
torch==2.0.1
transformers==4.30.2
sentence-transformers==2.2.2